disable bovonto random posts
