+++ bovonto.py
@@ -1,5 +1,32 @@
+import random, discord, asyncio
+from discord.ext import commands, tasks
+
+client = commands.Bot(command_prefix='!', description='Butterborg is online.', add=True)
+
+def pickRandomLine(name, icon, lines):
+    randLine = random.randint(0, len(lines) - 1)
+    line = lines[randLine]
+    e = discord.Embed(description=line)
+    e.set_author(name=name, icon_url=icon)
+    return e
+
+async def makePitch():
+    #Sets up status as ready
+    await client.wait_until_ready()
+    #Defines channels to send to (TOKENIZE THIS ASAP)
+    channelChoices = [465938202503413771,465991265557676054,644678362413006909,465950091044454411,466672962506981406]
+
+    #Loops through and chooses channels to send pitch to
+    while client.is_ready:
+        cycleChannels = random.randrange(len(channelChoices))
+        targetChannel = channelChoices[cycleChannels]
+        setChannel = client.get_channel(targetChannel)
+        await setChannel.send(embed=pickRandomLine(name='Bovonto Bot', icon= embedBovontoIcon, lines=pitches))
+        await asyncio.sleep(random.randrange(360,14400))
+
 embedBovontoIcon = "https://imgur.com/8aCQlV5.png"
-pitches = ["Hello, weary traveller! May I slake your thirst with a frosty Bovonto?",
+pitches = [
+"Hello, weary traveller! May I slake your thirst with a frosty Bovonto?",
 "Ah! In this summer heat a Bovonto will keep you cool!",
 "Good poster! May I interest you in a Bovonto this day?",
 "The pleasure of a Bovonto is nontrivial! Please, enjoy!",
@@ -22,5 +49,8 @@ pitches = ["Hello, weary traveller! May I slake your thirst with a frosty Bovont
 "I am a machine! To spread Bovonto's gospel, my purpose! Would you indulge me?",
 "Soldiers die. Forests burn. Bovonto Refreshes!",
 "Bovonto has been tied to fewer urinary tract infections than any other grape cola. Please enjoy!",
-"Weep not for the lost. Drink Bovonto in their name!"]
-
+"Weep not for the lost. Drink Bovonto in their name!",
+"Leap for joy, bliss is here! Bovonto!",
+"Drink deep from the bottle, and your deepest wish will be granted: Bovonto.",
+"The rock of plenty, the rock of faith: Bovonto"
+]
